Implement toolbox with deprecation tool
In order to support name deprecation without cluttering up the interface with a
load of buttons (one on each name to toggle its deprecation), use a tool design,
where you pick the deprecation tool, and then click on names to toggle their deprecation.

While a tool is active, all inline controls are hidden, and name editability is disabled.

Clicking on a button in the toolbar or pressing escape will clear the current tool and
return the editor to normal.