Improve startup time
The uncached spritesheet is no longer loaded twice, and the editor will no longer
wait for the uncached spritesheet to be loaded before enabling the interface, as
it is not needed until saving.
This does mean that prior to the spritesheet loading, you'll be editing while
seeing the cached (potentially outdated) spritesheet, which could result in making
unnecessary or incorrect edits, however the spritesheet will usually be loaded
by the time you actually go to change something, so that seems a worthwhile
trade-off to be able to see the editor a few seconds earlier.

This also theoretically enables image editing support for IE10 (and any other
browser that supports everything needed except Image.crossOrigin).