Implement block and protection checking (fixes #2)
Now checks that you are not blocked, and if you have the necessary rights to
edit the IDs and spritesheet, prior to enabling the editor. Otherwise, kicks you
out with a notification why (and the block reason, if applicable).

Also added a shortcut for when the editor is destroyed prior to being enabled,
as attempting to remove the contenteditable attribute (even when it isn't there),
is particularly slow.