Remove toolbar dropzone, make each section a dropzone
It's simpler to just drop the images directly into the section you want to place
them in, rather than into the toolbar dropzone and hope nearestSection() is correct.

This also allows auto-scroll to be implemented, as you no longer have to drag near
the edge of the screen (where the toolbar is) to drop something.