Implement improved save workflow
Saving is now done in the toolbar, rather than in a fullscreen dialog.
This allows continuing to view and edit while writing the summary, and allows
using a single-line input for the summary rather than a multi-line textarea,
which is more accurate to how the summary will be shown (and how it is seen in
a normal edit window), means autocomplete will work, and that the enter key can
be used to save.

Changed to using forced redraws rather than setImmediate for animations; and rAF
in the last few cases that needed setImmediate, so that the large setImmediate
polyfill could be removed.

Also fixed a bug with the dialog where the width and height wasn't set on first
panel opening, resulting in no scrollbars if the content overflows.