Remove ability to sort names
It's rare to need to change which image a name is associated with, and it can still be easily
done by cut & pasting the name, so it's not worth having the clutter from a handle on each
name.

If someone really needs to mass move names could always just add a move tool, which
would be useful for mass moving images to different sections as well.