Convert to OOUI
OOUI is bloated and totally overkill[1] for what we need (some simple button and input
styles), and honestly just gets in the way most of the time, but MW UI is deprecated and
would have to be replaced eventually. Might as well do it now and at least we get icons
and nicer select box styling.

[1] OOUI core is ~350KB, whereas MW UI buttons and input was a mere ~8KB. OOUI icons
are another ~280KB, but at least the editor doesn't have to wait for them to load before it is
usable.