Simplify dialog styling, improve edit conflicts further
Leave dialog sizing to flex rather than setting it manually, and use more flex to allow the
dialog content to continue to be scrollable, this allows the edit conflict textareas to
dynamically scale to take up the full height of the dialog, and improves responsiveness
when resizing the window.

Edit conflict panel now invalidates the names table when closed, so any changes made in
the edit conflict textarea does not overwrite changes made in the editor.

Dialog is now placed outside of the content area, so it is not overlapped by the sidebar,
tabs, and netbar.

Convert panel function to use a config parameter to better accommodate the increasing
options, add a callback for each panel to call when the dialog is closed by any panel, and
document panel function better.