Fix scroll position being reset on load sometimes
If detaching the documentation caused the page to be too short,
the page was scrolled back to the top, then the editor smoothly
scrolled back down to the top of the documentation when it was
reattached. Since you don't actually see the documentation detach
it made it look like the whole page just jumped for no reason.
Now store the original position before detaching and restore it
after reattaching. Then the editor can scroll if it actually needs
to.