Replace new/replaced images with updated sprites after saving
Previously new and replaced images would just stay there until you
reloaded the page, so they wouldn't be real sprites. Now the
spritesheet is overwritten with the local data that was just uploaded
(since we can't reliably use an image for awhile after its uploaded,
because cache), and the replaced images are swapped back to the
original sprite HTML. If new images were added, then the whole
documentation has to be refreshed from the server so it can parse
the new sprites.