Fix initial section renaming, and section highlighting
There always has to be one section, and trying to delete it should
force the text to be the default name. This was broken in part due
to the action being incorrect, so I added a console error for that,
and simplified how the check works so it works as if you had just
typed the default category name yourself.

Also made the section headline be inline-block so double/triple
click highlighting works correctly.