Cloudfront supports CORS now!
Also added the version query parameter to the URL, which seems to
trigger a cache miss in Cloudfront which should guarantee that it is
the absolute latest version from the server, meaning no more
overwriting previous spritesheet edits!

The versioned URLs also have the same 24 hour cache time as the
unversioned ones, so it's actually safe to use them in the stylesheet
to force an immediate (well at least ones the client's stylesheet
updates) spritesheet update for the whole wiki, allowing avoiding
any broken or missing sprites. Requires investigating how feasible
it would be to automatically update the stylesheet if an admin
makes the sprite edit, or embed the spritesheet URL in the page.