Fix content refreshing
The timestamp returned by `REVISIONTIMESTAMP`, uses the server
timezone, whereas the timestamp returned by the API is always UTC,
so the content request wouldn't accurately realise that the content
was changed on wiki's not using UTC. Additionally, changing refresh
to output the doc container broke actually inserting new content.

So a request is made to get the server's timezone offset so the
API timestamp can be corrected, and the entire doc container will
be replaced when refreshing (and reverting) the content. Since the
element `$doc` refers to now changes, anything referring to it had
to be moved to be per-session, rather than global.

Also since a new query request was going to be added, it was a good
time to take the two existing startup query requests, and merge all
three in to one, as well as convert all requests to use format
version 2 to simplify the responses.