Merge settings and IDs
I want to move to using versioned inline URLs, and by specifying the
URL in the settings, it doesn't have a performance impact from calling
Widget:FileUrl as it only needs to be called once per page. So now
the editor needs to have access to the settings in order to update
the version the URL is using, hence the need to merge IDs and
settings, which I wanted to do for ages anyway.

So now the editor is passed the settings as JSON, and it has an
actual lua table serializer so it can convert the entire data object
to a lua table rather than building it by hand.

Also remove some old IE9 compat and simplify the saving function.