Support inline versioned URLs
Since we'll be using versioned URLs now, the spritesheet already on
the page is almost guaranteed to be the most recent, so the editor
no longer needs to wait for the spritesheet request to load before
enabling, significantly speeding up startup time on large spritesheets.
A new uncached spritesheet still has to be downloaded, because we need
one with CORS headers, plus there's still the possibility that someone
uploaded a new version without using the editor, and thus didn't
update the version URL, so it's still good to force an uncached
version.

The version number actually shown in the diff won't be accurate though
as the lua table will always be re-created with a new version
number (the table is re-created and saved AFTER uploading the
spritesheet), rather than cached so to better ensure the version number
hasn't already been cached with the wrong version by someone else.

After saving, the editor will overwrite the spritesheet with the
newly uploaded version's URL with the version number that was saved
in the module, rather than leaving the blob, so that it is
pre-cached in cloudfront for better performance (yes, we're actually
trying to not avoid cache for once).

Also added the processing animation to the download button, since it
now has to wait for the spritesheet to download before it can actually
download an individual image.