Fix names being marked as duplicate when using find in Edge
This happens because Edge sends a blur event to an element when
it moves the find highlight away from it (regardless of if the
element is focused), and it doesn't even send a focus event to the
element it moved the highlight to. As such, it thinks the original
name was `undefined`, and thus tries to change the name to the same
is its current text, making it seem like a duplicate of itself.

I also added an additional protection against preventing saving
if something is erroneously marked as a duplicate, by not preventing
saving if there's only 1 name marked as a duplicate on the page.