Support passing through additional sprite data
Been wanting to allow additional data for various purposes for
awhile, but couldn't be bothered writing the interface for it. So
instead of blocking on that, just pass the data through so people
can add it to the module manually.

Need to implement de-duplicating sprite data for aliases now. Wasn't
really an issue before since the data was probably not much longer
than the sprite name itself, but now that you can have any arbitrary
amount of additional data, it makes a lot more sense.

Also `typeof null === 'object'` ouch.